[Feature][Flink] remove FlinkJobMetricsSummary
