+++ seatunnel-translation/seatunnel-translation-flink/seatunnel-translation-flink-common/src/main/java/org/apache/seatunnel/translation/flink/metric/FlinkJobMetricsSummary.java
@@ -23,10 +23,15 @@ import org.apache.seatunnel.common.utils.StringFormatUtils;
 
 import org.apache.flink.api.common.JobExecutionResult;
 
+import lombok.extern.slf4j.Slf4j;
+
 import java.time.Duration;
 import java.time.LocalDateTime;
+import java.util.HashMap;
+import java.util.Map;
 
-public final class FlinkJobMetricsSummary {
+@Slf4j
+public class FlinkJobMetricsSummary {
 
     private final JobExecutionResult jobExecutionResult;
 
@@ -34,13 +39,16 @@ public final class FlinkJobMetricsSummary {
 
     private final LocalDateTime jobEndTime;
 
-    FlinkJobMetricsSummary(
+    public FlinkJobMetricsSummary(
             JobExecutionResult jobExecutionResult,
             LocalDateTime jobStartTime,
             LocalDateTime jobEndTime) {
         this.jobExecutionResult = jobExecutionResult;
         this.jobStartTime = jobStartTime;
         this.jobEndTime = jobEndTime;
+        log.info(
+                "FlinkJobMetricsSummary created for job: {}",
+                jobExecutionResult != null ? jobExecutionResult.getJobID() : "null");
     }
 
     public static Builder builder() {
@@ -80,8 +88,84 @@ public final class FlinkJobMetricsSummary {
         }
     }
 
+    public Map<String, Object> getMetrics() {
+        Map<String, Object> metrics = new HashMap<>();
+
+        if (jobExecutionResult == null) {
+            log.warn("JobExecutionResult is null, cannot get metrics");
+            return metrics;
+        }
+
+        try {
+            Map<String, Object> accumulatorResults = jobExecutionResult.getAllAccumulatorResults();
+
+            for (Map.Entry<String, Object> entry : accumulatorResults.entrySet()) {
+                String key = entry.getKey();
+                Object value = entry.getValue();
+
+                if (value instanceof Number) {
+                    long longValue = ((Number) value).longValue();
+
+                    if (key.contains("SinkWriteCount")) {
+                        metrics.put(MetricNames.SINK_WRITE_COUNT, longValue);
+
+                    } else if (key.contains("SinkWriteBytes")) {
+                        metrics.put(MetricNames.SINK_WRITE_BYTES, longValue);
+
+                    } else if (key.contains("SourceReceivedCount")) {
+                        metrics.put(MetricNames.SOURCE_RECEIVED_COUNT, longValue);
+
+                    } else if (key.contains("SourceReceivedBytes")) {
+                        metrics.put(MetricNames.SOURCE_RECEIVED_BYTES, longValue);
+                    }
+                }
+            }
+        } catch (Exception e) {
+            log.warn("Failed to get metrics from accumulators: {}", e.getMessage(), e);
+        }
+
+        log.info("Retrieved metrics from accumulators: {}", metrics);
+        return metrics;
+    }
+
+    private long getCounterValue(Map<String, Object> metrics, String name, long defaultValue) {
+        Object value = metrics.get(name);
+        if (value == null) {
+            return defaultValue;
+        }
+
+        if (value instanceof Number) {
+            return ((Number) value).longValue();
+        }
+
+        try {
+            return Long.parseLong(value.toString());
+        } catch (NumberFormatException e) {
+            log.warn(
+                    "Failed to parse counter value: {} = {}, using default: {}",
+                    name,
+                    value,
+                    defaultValue);
+            return defaultValue;
+        }
+    }
+
     @Override
     public String toString() {
+        Map<String, Object> metrics = getMetrics();
+
+        long sourceReadCount = getCounterValue(metrics, MetricNames.SOURCE_RECEIVED_COUNT, 0L);
+        long sourceReadBytes = getCounterValue(metrics, MetricNames.SOURCE_RECEIVED_BYTES, 0L);
+        long sinkWriteCount = getCounterValue(metrics, MetricNames.SINK_WRITE_COUNT, 0L);
+        long sinkWriteBytes = getCounterValue(metrics, MetricNames.SINK_WRITE_BYTES, 0L);
+
+        log.info(
+                "Final metrics - sourceRead: {}, sourceBytes: {}, sinkWrite: {}, sinkBytes: {}",
+                sourceReadCount,
+                sourceReadBytes,
+                sinkWriteCount,
+                sinkWriteBytes);
+
         return StringFormatUtils.formatTable(
                 "Job Statistic Information",
                 "Start Time",
@@ -91,16 +175,12 @@ public final class FlinkJobMetricsSummary {
                 "Total Time(s)",
                 Duration.between(jobStartTime, jobEndTime).getSeconds(),
                 "Total Read Count",
-                jobExecutionResult
-                        .getAllAccumulatorResults()
-                        .get(MetricNames.SOURCE_RECEIVED_COUNT),
+                sourceReadCount,
                 "Total Write Count",
-                jobExecutionResult.getAllAccumulatorResults().get(MetricNames.SINK_WRITE_COUNT),
+                sinkWriteCount,
                 "Total Read Bytes",
-                jobExecutionResult
-                        .getAllAccumulatorResults()
-                        .get(MetricNames.SOURCE_RECEIVED_BYTES),
+                sourceReadBytes,
                 "Total Write Bytes",
-                jobExecutionResult.getAllAccumulatorResults().get(MetricNames.SINK_WRITE_BYTES));
+                sinkWriteBytes);
     }
 }
